# openapi.yaml
openapi: 3.0.3
info:
  title: Woliz Products API
  version: 2.0.0
  description: >
    API for managing the Woliz product catalog. The catalog uses a temporal
    model: a product keeps a stable identity, but its characteristics
    (price, format, packaging, SKU) evolve over time via variants.
    Products can belong to the global catalog (nanostoreId = null) or to a
    specific nanostore (private product). Includes user registration,
    login, and JWT-based authentication.
servers:
  - url: http://localhost:3000/api
security:
  - bearerAuth: []
paths:
...
  /partenaires/{partenaireId}/villes/{villeId}/zones:
    get:
      summary: List zones for a city
      operationId: listZones
      security: []
      parameters:
        - name: partenaireId
          in: path
          required: true
          schema:
            type: integer
        - name: villeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A list of zones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Zone'
        '404':
          description: Partner or city not found

    post:
      summary: Add a zone to a city
      operationId: createZone
      security: []
      parameters:
        - name: partenaireId
          in: path
          required: true
          schema:
            type: integer
        - name: villeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewZone'
      responses:
        '201':
          description: Zone created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Zone'
        '400':
          description: Invalid input
        '401':
          description: Missing or invalid token
        '404':
          description: Partner or city not found
...
  /partenaires/{partenaireId}/villes/{villeId}/zones/{zoneId}/toggle:
    patch:
      summary: Toggle a zone's active status
      operationId: toggleZone
      security: []
      parameters:
        - name: partenaireId
          in: path
          required: true
          schema:
            type: integer
        - name: villeId
          in: path
          required: true
          schema:
            type: integer
        - name: zoneId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Zone toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Zone'
        '404':
          description: Zone not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse' 
...
  /partenaires/{partenaireId}/villes/{villeId}/zones/{zoneId}:
    get:
      summary: Get a zone by id
      operationId: getZone
      security: []
      parameters:
        - name: partenaireId
          in: path
          required: true
          schema:
            type: integer
        - name: villeId
          in: path
          required: true
          schema:
            type: integer
        - name: zoneId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Zone found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Zone'
        '404':
          description: Partner, city or zone not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update a zone
      operationId: updateZone
      security: []
      parameters:
        - name: partenaireId
          in: path
          required: true
          schema:
            type: integer
        - name: villeId
          in: path
          required: true
          schema:
            type: integer
        - name: zoneId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewZone'
      responses:
        '200':
          description: Zone updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Zone'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Partner, city or zone not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete a zone
      operationId: deleteZone
      security: []
      parameters:
        - name: partenaireId
          in: path
          required: true
          schema:
            type: integer
        - name: villeId
          in: path
          required: true
          schema:
            type: integer
        - name: zoneId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Zone deleted
        '404':
          description: Partner, city or zone not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'              
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    NewZone:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
          example: "Agdal"
        endpoint:
          type: string
          nullable: true
          description: >
            Route/URL used to reach this zone's product feed, e.g.
            api/:fourniseur/:ville/:zone/product
          example: "api/woliz-distrib/rabat/agdal/product"
        status:
         type: string
         enum: [actif, inactif]
         default: "actif"  
    Zone:
      allOf:
        - $ref: '#/components/schemas/NewZone'
        - type: object
          properties:
            id:
              type: integer
              example: 1
            villeId:
              type: integer
              example: 1
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time

    # ---- Users / auth ----
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "email ou mot de passe incorrect"